djangochat/requirements.txt
aiohttp==3.8.3
aiohttp-rest-framework==0.0.6
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.3
daphne==3.0.2
dbus-python==1.3.2
Django==4.1.3
django-filter==22.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
Flask==2.2.2
frozenlist==1.3.3
greenlet==2.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
MouseInfo==0.1.3
multidict==6.0.3
mysqlclient==2.1.1
packaging==22.0
Pillow==9.3.0
plyer==2.1.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==22.1.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.6
pywhatkit==5.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.0b1
sqlparse==0.4.3
twilio==7.15.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wikipedia==1.4.0
yarl==1.8.2
zipp==3.11.0
zope.interface==5.5.2
